Добавить обработку ошибки декодирования

+++ engine/injection/controllers/data_project_controller.go
@@ -20,8 +20,12 @@ func (c *DataProjectController) AddDataProject(
 	writer http.ResponseWriter,
 	request *http.Request) {
 	var intent DataProject
-	Decode(request, &intent, writer)
-
+	decodeError := Decode(request, &intent)
+	if decodeError != nil {
+		writer.WriteHeader(http.StatusBadRequest)
+		writer.Write([]byte(decodeError.Error()))
+		return
+	}
 	_, err := c.Db.Exec("insert into project_config_data (data_id, project_config_id)values ($1, $2)", intent.DataId, intent.ProjectConfigId)
 	if err != nil {
 		writer.WriteHeader(http.StatusBadRequest)
@@ -61,8 +65,12 @@ func (c *DataProjectController) DeleteDataProject(
 	writer http.ResponseWriter,
 	request *http.Request) {
 	var intent DataProject
-	Decode(request, &intent, writer)
-
+	decodeError := Decode(request, &intent)
+	if decodeError != nil {
+		writer.WriteHeader(http.StatusBadRequest)
+		writer.Write([]byte(decodeError.Error()))
+		return
+	}
 	_, err := c.Db.Exec("delete from project_config_data where (data_id = $1 AND project_config_id = $2)", intent.DataId, intent.ProjectConfigId)
 	if err != nil {
 		writer.WriteHeader(http.StatusBadRequest)
